+++ CHANGELOG.md
@@ -6,9 +6,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - fix for output checking, custom verify functions are called just once
+- benchmarking returns multiple results not only time
 
 ### Added
 - support for kernels that use texture memory in CUDA
+- support for measuring energy consumption of CUDA kernels
 
 ## [0.2.0] - 2018-11-16
 ### Changed